Comment why we are checking for an empty string with escapeDigits
We need this due to how AGI parses the arguments passed to it.
If we just leave it empty it will assume that there is no arument there and the functions do not allow for this argument to be missing in their argc check.
See: https://github.com/asterisk/asterisk/blob/17.2/res/res_agi.c#L3955

Also add it to the other functions that could have no escapeDigits since the same would apply to them.